+++ spec/components/layout_footer_spec.rb
@@ -83,4 +83,73 @@ describe "Layout footer", type: :view do
       expect(link.attr('rel').to_s).to eq "noopener"
     end
   end
+
+  it "renders the footer meta links with attributes" do
+    render_component(
+      navigation: [
+        {
+          title: "Services and information",
+          columns: 2,
+          items: [
+            {
+              href: "/browse/benefits",
+              text: "Benefits",
+            },
+          ]
+        },
+      ],
+      meta: {
+        items: [
+          {
+            href: "/cymraeg",
+            text: "Rhestr o Wasanaethau Cymraeg",
+            attributes: {
+              lang: "cy",
+              hreflang: "cy"
+            },
+          },
+        ]
+      },
+    )
+
+    assert_select '.govuk-footer__meta-item a' do |link|
+      expect(link.attr('class').to_s).to eq "govuk-footer__link"
+      expect(link.attr('lang').to_s).to eq "cy"
+      expect(link.attr('hreflang').to_s).to eq "cy"
+    end
+  end
+
+  it "renders the footer meta link with attributes but will set the rel attribute if target is blank" do
+    render_component(
+      navigation: [
+        {
+          title: "Services and information",
+          columns: 2,
+          items: [
+            {
+              href: "/browse/benefits",
+              text: "Benefits",
+            },
+          ]
+        },
+      ],
+      meta: {
+        items: [
+          {
+            href: "/guidance/keeping-a-pet-pig-or-micropig",
+            text: "Keeping a pet pig or 'micropig'",
+            attributes: {
+              target: "_blank"
+            },
+          },
+        ]
+      },
+    )
+
+    assert_select '.govuk-footer__meta-item a' do |link|
+      expect(link.attr('class').to_s).to eq "govuk-footer__link"
+      expect(link.attr('target').to_s).to eq "_blank"
+      expect(link.attr('rel').to_s).to eq "noopener"
+    end
+  end
 end